chore: generify
