+++ ecc/bls24-315/fr/polynomial/polynomial_test.go
@@ -7,6 +7,9 @@ package polynomial
 
 import (
 	"github.com/consensys/gnark-crypto/ecc/bls24-315/fr"
+	"github.com/leanovate/gopter"
+	"github.com/leanovate/gopter/gen"
+	"github.com/leanovate/gopter/prop"
 	"github.com/stretchr/testify/assert"
 	"math/big"
 	"testing"
@@ -204,4 +207,48 @@ func TestPolynomialText(t *testing.T) {
 	p := Polynomial{one, negTwo, one}
 
 	assert.Equal(t, "X² - 2X + 1", p.Text(10))
+}
+
+func TestPrecomputeLagrange(t *testing.T) {
+
+	testForDomainSize := func(domainSize uint8) bool {
+		polys := computeLagrangeBasis(domainSize)
+
+		for l := uint8(0); l < domainSize; l++ {
+			for i := uint8(0); i < domainSize; i++ {
+				var I fr.Element
+				I.SetUint64(uint64(i))
+				y := polys[l].Eval(&I)
+
+				if i == l && !y.IsOne() || i != l && !y.IsZero() {
+					t.Errorf("domainSize = %d: p_%d(%d) = %s", domainSize, l, i, y.Text(10))
+					return false
+				}
+			}
+		}
+		return true
+	}
+
+	t.Parallel()
+	parameters := gopter.DefaultTestParameters()
+
+	const maxLagrangeDomainSize = 12
+
+	parameters.MinSuccessfulTests = maxLagrangeDomainSize
+
+	properties := gopter.NewProperties(parameters)
+
+	properties.Property("l'th lagrange polynomials must evaluate to 1 on l and 0 on other values in the domain", prop.ForAll(
+		testForDomainSize,
+		gen.UInt8Range(2, maxLagrangeDomainSize),
+	))
+
+	properties.TestingRun(t, gopter.ConsoleReporter(false))
+}
+
+func TestLagrangeCache(t *testing.T) {
+	for _, i := range []int{5, 2, 8, 4, 6, 3, 0} {
+		b := getLagrangeBasis(uint8(i))
+		assert.Equal(t, b, getLagrangeBasis(uint8(i))) // second call must yield the same result
+	}
 }
