+++ internal/generator/polynomial/template/multilin.test.go.tmpl
@@ -1,8 +1,5 @@
 import (
 	"{{.FieldPackagePath}}"
-	"github.com/leanovate/gopter"
-	"github.com/leanovate/gopter/gen"
-	"github.com/leanovate/gopter/prop"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -51,41 +48,6 @@ func TestFoldBilinear(t *testing.T) {
 			t.Fail()
 		}
 	}
-}
-
-func TestPrecomputeLagrange(t *testing.T) {
-
-	testForDomainSize := func(domainSize uint8) bool {
-		polys := computeLagrangeBasis(domainSize)
-
-		for l := uint8(0); l < domainSize; l++ {
-			for i := uint8(0); i < domainSize; i++ {
-				var I {{.ElementType}}
-				I.SetUint64(uint64(i))
-				y := polys[l].Eval(&I)
-
-				if i == l && !y.IsOne() || i != l && !y.IsZero() {
-					t.Errorf("domainSize = %d: p_%d(%d) = %s", domainSize, l, i, y.Text(10))
-					return false
-				}
-			}
-		}
-		return true
-	}
-
-	t.Parallel()
-	parameters := gopter.DefaultTestParameters()
-
-	parameters.MinSuccessfulTests = int(maxLagrangeDomainSize)
-
-	properties := gopter.NewProperties(parameters)
-
-	properties.Property("l'th lagrange polynomials must evaluate to 1 on l and 0 on other values in the domain", prop.ForAll(
-		testForDomainSize,
-		gen.UInt8Range(2, maxLagrangeDomainSize),
-	))
-
-	properties.TestingRun(t, gopter.ConsoleReporter(false))
 }
 
 // TODO: Benchmark folding? Algorithms is pretty straightforward; unless we want to measure how well memory management is working
